+++ internal/pkg/runtime/engine/singularity/cleanup_linux.go
@@ -16,12 +16,12 @@ import (
 
 	"github.com/sylabs/singularity/internal/pkg/instance"
 	fakerootConfig "github.com/sylabs/singularity/internal/pkg/runtime/engine/fakeroot/config"
+	"github.com/sylabs/singularity/internal/pkg/util/crypt"
 	"github.com/sylabs/singularity/internal/pkg/util/priv"
 	"github.com/sylabs/singularity/internal/pkg/util/starter"
 	"github.com/sylabs/singularity/pkg/runtime/engine/config"
 	"github.com/sylabs/singularity/pkg/sylog"
 	"github.com/sylabs/singularity/pkg/util/capabilities"
-	"github.com/sylabs/singularity/pkg/util/crypt"
 )
 
 // CleanupContainer is called from master after the MonitorContainer returns.
