+++ internal/pkg/runtime/engine/singularity/rpc/server/server_linux.go
@@ -15,12 +15,12 @@ import (
 	"syscall"
 
 	args "github.com/sylabs/singularity/internal/pkg/runtime/engine/singularity/rpc"
+	"github.com/sylabs/singularity/internal/pkg/util/crypt"
 	"github.com/sylabs/singularity/internal/pkg/util/fs"
 	"github.com/sylabs/singularity/internal/pkg/util/mainthread"
 	"github.com/sylabs/singularity/internal/pkg/util/user"
 	"github.com/sylabs/singularity/pkg/sylog"
 	"github.com/sylabs/singularity/pkg/util/capabilities"
-	"github.com/sylabs/singularity/pkg/util/crypt"
 	"github.com/sylabs/singularity/pkg/util/loop"
 	"github.com/sylabs/singularity/pkg/util/namespaces"
 	"golang.org/x/sys/unix"
